UE5 fixes on MacOS
Delegate prototype switched to double from float.
PlatformData changed to GetPlatformData().
Added section for TargetPlatform Mac for libassimp.dylib.
Changed enum from Point to PointLight to avoid conflict.

+++ Source/UE_Assimp/Public/AssimpFunctionLibrary.h
@@ -18,8 +18,8 @@
 
 
 
-DECLARE_DYNAMIC_DELEGATE_TwoParams(FOnProgressUpdated,float,NormalPercentage,UAIScene* ,ImportedScene);
-DECLARE_DYNAMIC_DELEGATE_TwoParams(FOnImportSceneComplete,const TArray<UAIScene*>&,ImportedScenes,float ,TotalTimeElapsed);
+DECLARE_DYNAMIC_DELEGATE_TwoParams(FOnProgressUpdated,double,NormalPercentage,UAIScene* ,ImportedScene);
+DECLARE_DYNAMIC_DELEGATE_TwoParams(FOnImportSceneComplete,const TArray<UAIScene*>&,ImportedScenes,double ,TotalTimeElapsed);
 UCLASS()
 class UE_ASSIMP_API UAssimpFunctionLibrary : public UBlueprintFunctionLibrary
 {
